+++ src/simulation/simulation.jl
@@ -1,12 +1,31 @@
 using Random
 
-function simulate!(model::Model, time_vector::Vector{Float64})
+function simulate!(
+        model::Model, time_vector::Vector{Float64};
+        host_samples_population::Dict{String, Int64}=Dict{String, Int64}())
+
     # Interventions
     interventions = sort(model.interventions, by=intervention -> intervention.time)
     intervention_tracker = 0 # keeps track of what the next intervention should be
 
     # History
-    his_tracker = 0
+    his_tracker = 1
+    compartment_vars = Dict{String, Matrix{Int64}}()
+    for p in model.populations
+        compartment_vars[p.id] = Matrix{Int64}(
+            undef, NUM_COMPARTMENTS, length(time_vector)
+        )
+        # tracks uninfected naive, infected naive, uninfected immune, infected immune,
+        # and dead for this population at each time point
+    end
+    host_samples = Dict{String, Matrix{StaticHost}}()
+    for p in model.populations
+        if haskey(host_samples_population, p.id)
+            host_samples[p.id] = Matrix{StaticHost}(
+                undef, host_samples_population[p.id], length(time_vector)
+            )
+        end
+    end
 
     # Gillespie
     time = 0.0
@@ -55,9 +74,20 @@ function simulate!(model::Model, time_vector::Vector{Float64})
                 time = time_vector[end]
             end
         end
+        while his_tracker <= length(time_vector) && time >= time_vector[his_tracker]
+            for p in model.populations
+                compartment_vars[p.id][:,his_tracker] = p.compartment_vars
+                if haskey(host_samples_population, p.id)
+                    for i in 1:min(host_samples_population[p.id], length(p.hosts))
+                        host_samples[p.id][i,his_tracker] = staticHost(p.hosts[i])
+                    end
+                end
+            end
+            his_tracker += 1
+        end
     end
 
     println(evt_count)
 
-    # return model
+    return model, compartment_vars, host_samples
 end